rename: 멤버, 시큐리티 폴더 정리
