+++ src/main/java/com/kosa/fillinv/global/security/filter/JwtAuthenticationFilter.java
@@ -1,5 +1,7 @@
-package com.kosa.fillinv.global.security;
+package com.kosa.fillinv.global.security.filter;
 
+import com.kosa.fillinv.global.security.jwt.JWTUtil;
+import com.kosa.fillinv.global.security.details.CustomMemberDetails;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.kosa.fillinv.global.response.ErrorCode;
 import com.kosa.fillinv.global.response.ErrorResponse;
