+++ src/main/java/com/kosa/fillinv/member/dto/member/SignUpDto.java
@@ -1,4 +1,4 @@
-package com.kosa.fillinv.member.dto;
+package com.kosa.fillinv.member.dto.member;
 
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
