+++ src/main/java/com/kosa/fillinv/member/dto/profile/ProfileImageRequestDto.java
@@ -1,4 +1,4 @@
-package com.kosa.fillinv.member.dto;
+package com.kosa.fillinv.member.dto.profile;
 
 import lombok.Getter;
 import lombok.NoArgsConstructor;
