+++ src/main/java/com/kosa/fillinv/member/dto/profile/ProfileResponseDto.java
@@ -1,4 +1,4 @@
-package com.kosa.fillinv.member.dto;
+package com.kosa.fillinv.member.dto.profile;
 
 import com.kosa.fillinv.category.dto.CategoryResponseDto;
 import lombok.Builder;
