+++ src/main/java/com/kosa/fillinv/member/dto/security/LoginRequest.java
@@ -1,4 +1,4 @@
-package com.kosa.fillinv.member.dto;
+package com.kosa.fillinv.member.dto.security;
 
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
