+++ src/main/java/com/kosa/fillinv/member/service/MemberService.java
@@ -5,10 +5,10 @@ import com.kosa.fillinv.category.entity.Category;
 import com.kosa.fillinv.category.exception.CategoryException;
 import com.kosa.fillinv.category.repository.CategoryRepository;
 import com.kosa.fillinv.global.response.ErrorCode;
-import com.kosa.fillinv.member.dto.IntroductionRequestDto;
+import com.kosa.fillinv.member.dto.profile.IntroductionRequestDto;
 import com.kosa.fillinv.member.exception.MemberException;
-import com.kosa.fillinv.member.dto.ProfileResponseDto;
-import com.kosa.fillinv.member.dto.SignUpDto;
+import com.kosa.fillinv.member.dto.profile.ProfileResponseDto;
+import com.kosa.fillinv.member.dto.member.SignUpDto;
 import com.kosa.fillinv.member.entity.Member;
 import com.kosa.fillinv.member.entity.Profile;
 import com.kosa.fillinv.member.repository.MemberRepository;
